versao 5.1
fonte arial incluida, informando tiros disparados e energia armazenada;
zombies nascendo de 1 em 1(arrumar ainda)

+++ objects.h
@@ -36,7 +36,7 @@ struct Zombies
     bool live; // Define quando um novo Zombie surge
     int life;
     float speed; // Velocidade do Zombie
-    float boundx; // Colisao X
+    int boundx; // Colisao X
     int boundy; // Colisao Y
 
 };
